Use FormulaParser for differentiate
